+++ client/ui/components/header/HeaderProfile.jsx
@@ -3,8 +3,8 @@ import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
 import {Meteor} from "meteor/meteor";
 import {useTracker} from "meteor/react-meteor-data";
 import {Link} from "react-router-dom";
-import {PROFILE_PUBLICATION} from "../../../../imports/modules/profile/enums/publication.js";
-import {profileRepository} from "../../../../imports/modules/profile/profileRepository.js";
+import {PROFILE_PUBLICATION} from "../../../../imports/modules/profiles/enums/publication.js";
+import {profileRepository} from "../../../../imports/modules/profiles/profileRepository.js";
 import {ROUTE} from "../../../routes/enums/route.js";
 
 export const HeaderProfile = () => {
