+++ client/ui/pages/auth/password/Register.jsx
@@ -2,7 +2,7 @@ import React, {useRef} from 'react';
 import {Button, Label, TextInput} from 'flowbite-react';
 import {STATE_AUTH_PASSWORD_FORM} from "./enums/state.js";
 import {Accounts} from "meteor/accounts-base";
-import {profileInsert} from "../../../../../imports/modules/profile/profile.methods.js";
+import {profileInsert} from "../../../../../imports/modules/profiles/profile.methods.js";
 
 export const Register = ({onStateChange}) => {
   const emailRef = useRef();
