+++ server/modules.js
@@ -1,7 +1,7 @@
 import '../imports/modules/dummy/server';
 import '../imports/modules/links/server';
-import '../imports/modules/profile/server';
-import '../imports/modules/user/server';
+import '../imports/modules/profiles/server';
+import '../imports/modules/users/server';
 import '../imports/modules/userPartition/server';
-import '../imports/modules/userRole/server';
+import '../imports/modules/userRoles/server';
 import '../imports/modules/tickets/server';
